+++ NEWS.rst
@@ -7,6 +7,7 @@ of every change, see the Git log.
 Latest
 ------
 * tbd
+* Minor: Enabled the kodoc_binary16 finite field.
 * Minor: Updated the license to STEINWURF EVALUATION LICENSE 1.0.
 
 12.0.0